+++ src/main/java/org/fsb/municipalite/controllers/ProjectPageController.java
@@ -4,12 +4,19 @@ import javafx.beans.binding.Bindings;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
+import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.scene.control.*;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.scene.image.Image;
+import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.Pane;
+import javafx.stage.Stage;
+import javafx.stage.StageStyle;
+
 import org.fsb.municipalite.entities.Equipe;
 import org.fsb.municipalite.entities.Projet;
 import org.fsb.municipalite.services.impl.EquipeServiceImpl;
@@ -32,11 +39,14 @@ public class ProjectPageController implements Initializable {
     @FXML TableColumn<Projet,Long> version ;
     @FXML TableColumn<Projet,String> place;
     @FXML TableColumn<Projet, Long> teamColumn;
-    public ObservableList data;
+    public ObservableList<Projet> data;
     @FXML
     TextField searchBox;
 
-
+    //define your offsets here
+    private double xOffset = 0;
+    private double yOffset = 0;
+    
     @Override
     public void initialize(URL location, ResourceBundle resources) {
         tableview.getSelectionModel().setSelectionMode(SelectionMode.MULTIPLE);
@@ -47,23 +57,15 @@ public class ProjectPageController implements Initializable {
         budget.setCellValueFactory(new PropertyValueFactory<Projet, Integer>("budget"));
         version.setCellValueFactory(new PropertyValueFactory<Projet,Long>("version"));
         place.setCellValueFactory(new PropertyValueFactory<Projet,String>("lieu"));
-        teamColumn.setCellValueFactory(new PropertyValueFactory<Projet,Long>("equipe"));
+        teamColumn.setCellValueFactory(new PropertyValueFactory<Projet,Long>("equipeValue"));
         data  =  FXCollections.observableArrayList();
         ProjetServiceImpl projetService = new ProjetServiceImpl();
         List<Projet> list = projetService.selectAll();
 
         for(Projet p : list) {
-           /* ObservableList row  =  FXCollections.observableArrayList();
-            row.add(p.getId());
-            row.add(p.getName());
-            row.add(p.getDateDebut());
-            row.add(p.getDateFin());
-            row.add(p.getBudget());
-            row.add(p.getVersion());
-            row.add(p.getLieu());
-            row.add(p.getEquipe().getId());*/
             data.add(p);
         }
+        
         tableview.setItems(data);
 
         searchBox.textProperty().addListener((observable, oldValue, newValue) -> {
@@ -128,12 +130,36 @@ public class ProjectPageController implements Initializable {
             FXMLLoader f = new FXMLLoader();
             f.setLocation(getClass().getResource("/interfaces/ProjectAdd.fxml"));
             Pane projetDialogPane = f.load();
-            ProjectAddController pac = f.getController();
+            ProjectDialogController pac = f.getController();
 
             Dialog<ButtonType> d = new Dialog<>();
+            
+            //this is just for adding an icon to the dialog pane
+			Stage stage = (Stage) d.getDialogPane().getScene().getWindow();
+			stage.getIcons().add(new Image("/assets/img/icon.png"));
+			
             d.setDialogPane((DialogPane) projetDialogPane);
             d.setTitle("add project");
-
+            d.setResizable(false);
+			d.initStyle(StageStyle.UNDECORATED);
+			
+			//these two are for moving the window with the mouse
+			projetDialogPane.setOnMousePressed(new EventHandler<MouseEvent>() {
+	           @Override
+	           public void handle(MouseEvent event) {
+	               xOffset = event.getSceneX();
+	               yOffset = event.getSceneY();
+	           }
+			});
+       
+			projetDialogPane.setOnMouseDragged(new EventHandler<MouseEvent>() {
+	           @Override
+	           public void handle(MouseEvent event) {
+	               d.setX(event.getScreenX() - xOffset);
+	               d.setY(event.getScreenY() - yOffset);
+	           }
+            });
+            
             //name field listener
             pac.name.textProperty().addListener((observable, oldValue, newValue) -> {
                 if(!isAlpha(newValue)) {
@@ -165,16 +191,28 @@ public class ProjectPageController implements Initializable {
             });
 
             pac.end.valueProperty().addListener((observable, oldValue, newValue) -> {
-                if(newValue != null) {
-                    if(newValue.isBefore(pac.start.getValue())) {
-                        pac.warning.setVisible(true);
+            	try {
+            		if(newValue != null) {
+                        if(newValue.isBefore(pac.start.getValue())) {
+                            pac.warning.setVisible(true);
+                        }
+                        else {
+                            pac.warning.setVisible(false);
+                        }
                     }
-                    else {
-                        pac.warning.setVisible(false);
-                    }
-                }
+            	}catch(Exception e) {
+            		System.out.println("end date picked");
+            	}
+                
             });
 
+            //to apply css on the dialog pane buttons
+			d.getDialogPane().lookupButton(ButtonType.APPLY).getStyleClass().add("dialogButtons");
+			d.getDialogPane().lookupButton(ButtonType.CANCEL).getStyleClass().add("dialogButtons");
+			
+			//make name field first to be selected
+			pac.name.requestFocus();
+            
             d.getDialogPane().lookupButton(ButtonType.APPLY).disableProperty().bind(Bindings.createBooleanBinding(() ->
                             pac.name.getText().isEmpty()
                                     || pac.budget.getText().isEmpty()
@@ -182,11 +220,13 @@ public class ProjectPageController implements Initializable {
                                     ||!isNumeric(pac.budget.getText())
                                     || pac.start.getValue() == null
                                     || pac.end.getValue() == null
-                                    ||pac.start.getValue().isAfter(pac.end.getValue()),
+                                    || pac.start.getValue().isAfter(pac.end.getValue())
+                                    || pac.team.getValue() == null,
                                     pac.name.textProperty(),
                                     pac.budget.textProperty(),
                                     pac.start.valueProperty(),
-                                    pac.end.valueProperty()));
+                                    pac.end.valueProperty(),
+            						pac.team.valueProperty()));
 
 
             Optional<ButtonType> clickedButton = d.showAndWait();
@@ -216,22 +256,34 @@ public class ProjectPageController implements Initializable {
     @FXML
     public void remove(ActionEvent event) {
         if(tableview.getSelectionModel().getSelectedItem() != null) {
-            ObservableList<Projet> selectedItems = tableview.getSelectionModel().getSelectedItems();
-            for (Projet p : selectedItems){
-                ProjetServiceImpl projetService = new ProjetServiceImpl();
-                projetService.remove(p.getId());
-            }
+        	
+        	Alert alert = new Alert(AlertType.CONFIRMATION);
+			Stage stage = (Stage) alert.getDialogPane().getScene().getWindow();
+			//this is just for adding an icon to the dialog pane
+			stage.getIcons().add(new Image("/assets/img/icon.png"));
+			alert.setTitle("Delete Project ?");
+			alert.setHeaderText(null);
+			alert.setContentText("Are you Sure You Want to Delete Selected Item(s) ?");
+			Optional <ButtonType> action = alert.showAndWait();
+			if(action.get() == ButtonType.OK) {
+	            ObservableList<Projet> selectedItems = tableview.getSelectionModel().getSelectedItems();
+	            for (Projet p : selectedItems){
+	                ProjetServiceImpl projetService = new ProjetServiceImpl();
+	                projetService.remove(p.getId());
+	            }
+	            refresh(event);
+			}
         }
-        refresh(event);
+        
     }
 
     @FXML
     public void update(ActionEvent event) {
         try {
             FXMLLoader f = new FXMLLoader();
-            f.setLocation(getClass().getResource("/interfaces/ProjectUpdate.fxml"));
-            Pane materielDialogPane = f.load();
-            ProjectUpdateController puc = f.getController();
+            f.setLocation(getClass().getResource("/interfaces/ProjectAdd.fxml"));
+            Pane projectDialogPane = f.load();
+            ProjectDialogController puc = f.getController();
 
             if(tableview.getSelectionModel().getSelectedItem() != null) {
 
@@ -240,10 +292,33 @@ public class ProjectPageController implements Initializable {
                 Projet projet = projetService.getById(m.getId());
 
                 puc.setProjectDialogPane(projet);
+                puc.titleLabel.setText("Update Project");
                 Dialog<ButtonType> d = new Dialog<>();
-                d.setDialogPane((DialogPane) materielDialogPane);
+                Stage stage = (Stage) d.getDialogPane().getScene().getWindow();
+				stage.getIcons().add(new Image("/assets/img/icon.png"));
+				
+                d.setDialogPane((DialogPane) projectDialogPane);
                 d.setTitle("Update projet");
-
+                d.setResizable(false);
+				d.initStyle(StageStyle.UNDECORATED);
+				
+				//these two are for moving the window with the mouse
+				projectDialogPane.setOnMousePressed(new EventHandler<MouseEvent>() {
+		           @Override
+		           public void handle(MouseEvent event) {
+		               xOffset = event.getSceneX();
+		               yOffset = event.getSceneY();
+		           }
+				});
+	       
+				projectDialogPane.setOnMouseDragged(new EventHandler<MouseEvent>() {
+		           @Override
+		           public void handle(MouseEvent event) {
+		               d.setX(event.getScreenX() - xOffset);
+		               d.setY(event.getScreenY() - yOffset);
+		           }
+	            });
+				
                 //name field listener
                 puc.name.textProperty().addListener((observable, oldValue, newValue) -> {
                     if(!isAlpha(newValue)) {
@@ -280,7 +355,13 @@ public class ProjectPageController implements Initializable {
                         }
                     }
                 });
-
+                //to apply css on the dialog pane buttons
+				d.getDialogPane().lookupButton(ButtonType.APPLY).getStyleClass().add("dialogButtons");
+				d.getDialogPane().lookupButton(ButtonType.CANCEL).getStyleClass().add("dialogButtons");
+				
+				//make name field first to be selected
+				puc.name.requestFocus();
+                
                 d.getDialogPane().lookupButton(ButtonType.APPLY).disableProperty().bind(Bindings.createBooleanBinding(() ->
                                 puc.name.getText().isEmpty()
                                         || puc.budget.getText().isEmpty()
@@ -311,7 +392,7 @@ public class ProjectPageController implements Initializable {
 
     //Test if String is alphabetical
     public boolean isAlpha(String name) {
-        return name.matches("[a-zA-Z]+");
+        return name.matches("[a-zA-Z ]+");
     }
 
 }
